helpers: move sha precompute to sha file
