+++ packages/helpers/src/shaHash.ts
@@ -1,9 +1,78 @@
-// import { createHash } from "crypto";
 import * as CryptoJS from 'crypto';
-import { assert, int64toBytes, int8toBytes, mergeUInt8Arrays, Uint8ArrayToHex } from "./binaryFormat";
-// const { webcrypto, KeyObject } = await import('crypto');
-// const { subtle } = webcrypto;
+import { assert, int64toBytes, int8toBytes, mergeUInt8Arrays } from "./binaryFormat";
 import { Hash } from "./fast-sha256";
+
+export function findIndexInUint8Array(
+  array: Uint8Array,
+  selector: Uint8Array
+): number {
+  let i = 0;
+  let j = 0;
+  while (i < array.length) {
+    if (array[i] === selector[j]) {
+      j++;
+      if (j === selector.length) {
+        return i - j + 1;
+      }
+    } else {
+      j = 0;
+    }
+    i++;
+  }
+  return -1;
+}
+
+export function padUint8ArrayWithZeros(array: Uint8Array, length: number) {
+  while (array.length < length) {
+    array = mergeUInt8Arrays(array, int8toBytes(0));
+  }
+  return array;
+}
+
+export function generatePartialSHA({
+  body,
+  bodyLength,
+  selectorString, // String to split the body
+  maxRemainingBodyLength, // Maximum allowed length of the body after the selector
+}: {
+  body: Uint8Array;
+  bodyLength: number;
+  selectorString?: string;
+  maxRemainingBodyLength: number;
+}) {
+  let selectorIndex = 0;
+
+  // TODO: See if this (no preselector) could be handled at the circuit level
+  if (selectorString) {
+    const selector = new TextEncoder().encode(selectorString);
+    selectorIndex = findIndexInUint8Array(body, selector);
+  }
+
+  const shaCutoffIndex = Math.floor(selectorIndex / 64) * 64;
+  const precomputeText = body.slice(0, shaCutoffIndex);
+  let bodyRemaining = body.slice(shaCutoffIndex);
+
+  const bodyRemainingLength = bodyLength - precomputeText.length;
+
+  if (bodyRemainingLength > maxRemainingBodyLength) {
+    throw new Error(
+      `Remaining body ${bodyRemainingLength} after the selector is longer than max (${maxRemainingBodyLength})`
+    );
+  }
+
+  if (bodyRemaining.length % 64 !== 0) {
+    throw new Error(`Remaining body was not padded correctly with int64s`);
+  }
+
+  bodyRemaining = padUint8ArrayWithZeros(bodyRemaining, maxRemainingBodyLength);
+  const precomputedSha = partialSha(precomputeText, shaCutoffIndex);
+
+  return {
+    precomputedSha,
+    bodyRemaining,
+    bodyRemainingLength,
+  };
+}
 
 export function shaHash(str: Uint8Array) {
   return CryptoJS.createHash('sha256').update(str).digest();
